+++ packages/auto_mappr/CHANGELOG.md
@@ -1,5 +1,8 @@
 [//]: # (## Unreleased)
 
+- Improve error messages. [#147](https://github.com/netglade/auto_mappr/pull/147)
+- When the source is not null, `TypeConverter<Object, Object>` and `TypeConverter<Object, Object?>` now support mapping of source field `Object?` -> target field `Object?`. [#142](https://github.com/netglade/auto_mappr/pull/142)
+
 ## 2.1.1
 - Fix auto_mappr_annotation dependency. [#130](https://github.com/netglade/auto_mappr/pull/131)